Fix "ld (i16), (i8h)" microcode
Previously this tried to do "YO" and "P+" at the same time, which is not possible because "P+"
reuses one of the ALU control bits. Simultaneously, a uasm bug failed to detect the problem,
so the previous microcode would always *execute* the "i16" argument right after the load. Normally
this was a pseudo-register, so the "i16" argument's opcode was 0xff, which is slownop, so it didn't
cause any obvious problem, except a performance degradation.