kilo: faster waitreadbyte()
This is still not quite fast enough to consume arrow keys at 115200 baud.
I think for a 1 MHz clock frequency, we need to spend fewer than 86 cycles
per byte in order to not miss any. Currently waitreadbyte() looks like it
takes about 52 cycles in the worst case, but there is overhead in readkey().

It should be doable if the critical path of escape sequence handling is all
moved to assembly language.