blkdev: stop thinking we're out of disk space
Previously the kernel would eventually conclude that it had run out of disk
space, because it started searching at the previous `freeblkgroup` offset for
each block of the free space bitmap, without wrapping around once reaching
the end. For example if `freeblkgroup` got to 255, and every block of the
free space bitmap had the last place fully used, it would think the disk was
full.

The fix is to wrap back around to 0.