kilo: fix use of uninitialised value
Previously this worked by accident because the inefficient local variable
initialisation meant that the 0 (initialising x) was pushed on to the stack
and then popped off again. With the recent improvements, this no longer happens,
so i was initialised to random garbage instead.

Interestingly, 2 different commits both exposed the broken behaviour ( ad2cd2e
and 5d282b8 ), which made it quite hard to work out what was wrong: I git bisected
to find the offending commit, reverted it, and the bug remained!