kernel: don't crash when restoring a large program into memory
The problem was that in the event that you ask read() for more than 254 words,
and the current pointer is at the start of a block on disk, read() will ask cf_blkread()
to read directly into the user buffer, rather than memcpy()ing to it.

The problem is that if the program we're restoring is very large, this can lead to
cf_blkread() putting 254 words into memory, even when the remaining words in the file
is lower than that, which overwrites some of the kernel causing a crash.

In particular, this case is triggered by typing ^Z to spawn a subshell. The kernel
doesn't know how much memory the running program is using, so it saves it all,
which would cause a crash when trying to restore it.

This fixes the bug identified at 35:04 in https://www.youtube.com/watch?v=DEs3pz48LUQ&t=35m4s