fs_write: don't corrupt the disk
There are 2 changes here:

1. we no longer read the block header off disk if the block number matches
   what we already have in the buffer

2. we write a new block's header out to disk immediately instead of only
   leaving it in the buffer

Either of these changes alone would fix the disk corruption problem, which was
that if the user program has specified a custom buffer, then we can allocate
a block for the file, write its number into the buffer, but then throw that
information away by reading the existing block contents off disk.

The most-performant option would be to stick with only change no. 1, to reduce
the amount of time wasted on writing blocks to disk, but there is still a risk
that the user program crashes the computer before the new block's header ever
gets written, which would result in the same kind of disk corruption still
happening.

So I consider change no. 2 to be required, but change no. 1 is still helpful
for the slight performance benefit of not having to read in the block headers.